feat: Add day & night style function
First, by adding the overall style of the app and adding day & night mode to the menu item, it was set to change the background color of the app according to the selected mode. And it saves the state of the app so that it can always maintain the state even when the screen returns or goes back.
Function used: onCreateOptionMenu(), onOptionItemSelected() and onSaveInstanceState().