feat: Add notification upadate & cancel function
You can cancel the notification by giving the notification ID as an argument to the cancel() function. After converting a specific image to a bitmap, it can be updated after applying it as a notification style. The state of the button is set after each function as a boolean so that the button can be clicked only at a specific moment.