feat: Add NotifiacationJobService class
This class extends JobService.

First, createNotificationChannel() initializes NotificationManger and initializes the notification channel by checking if the SDK version is higher than OREO. And it creates a notification channel.

In onStartJob(), a notification channel is created by calling createNotification(), and a Pending Intent that can come to the app's MainActivity when the notification is clicked is created, and a notification is set and a notification is made using a builder.